requirements.txt
python_dateutil==2.8.2
selenium==4.8.2
